Remove JMimeMagic
Fix type issues
Update RAFReader to use a MappedByteBuffer
Add method for recognizing JSON
General code cleaning

+++ src/main/java/no/stelar7/cdragon/wad/data/content/WADContentHeaderV1.java
@@ -6,7 +6,7 @@ import lombok.*;
 public class WADContentHeaderV1
 {
     protected long    pathHash;
-    protected long    offset;
+    protected int     offset;
     protected int     compressedFileSize;
     protected int     fileSize;
     protected boolean compressed;
